Major refactoring of program. Separated files into modules based on their responsibilities. Created the PokemonRouteSearch directory to handle the main logic and user interaction. Created PokeAPI_wrapper for the API wrapper and the objects created from those responses.
Moved the objects created from query responses to response_models.py and moved all API related functions to api_wrapper.py.

Made a new ApiWrapper class in api_wrapper.py and made LocationPokemonSearch and PokemonLocationSearch subclasses of that class.
